pythonProject2/requirements.txt
Flask~=2.0.2
Werkzeug~=2.0.2
Pillow~=8.4.0
opencv-python~=4.5.4.60
numpy~=1.21.5
torch~=1.10.1
torchvision~=0.11.2
matplotlib~=3.5.1